requirements.txt
astroid @ file:///tmp/build/80754af9/astroid_1592495913683/work
brotlipy==0.7.0
certifi==2020.11.8
cffi==1.14.0
chardet @ file:///tmp/build/80754af9/chardet_1605303175790/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544480695/work
Django==2.2.5
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.18
mysqlclient @ file:///tmp/build/80754af9/mysqlclient_1594308743189/work
protobuf==3.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pylint @ file:///tmp/build/80754af9/pylint_1598624022510/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
pytz==2020.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
six @ file:///tmp/build/80754af9/six_1605205335545/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
typed-ast==1.4.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wrapt==1.11.2
